backend/requirements.txt
# Core
dj-rest-auth[with_social]~=5.0.1
Django~=4.2.7
django-bulk-sync~=3.3.0
django-cachalot~=2.6.1
django-cors-headers~=4.3.0
django-debug-toolbar~=4.2.0
django-extensions~=3.2.3
django-filter~=23.3
django-money~=3.4.0
django-phonenumber-field~=7.2.0
django-redis~=5.4.0
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.0
drf-yasg==1.21.7
joblib~=1.3.2
phonenumbers~=8.13.24
psycopg2~=2.9.9
redis~=5.0.1
rest-framework-roles~=1.0.3
